Created ServiceRequest super
Added a ServiceRequest super to make dealing with different types of service requests in the same functions easier
